# src/main/resources/META-INF/openapi.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "ToDo",
        "version": "1.0.0",
        "description": "Simple ToDo list app"
    },
    "paths": {
        "/todos": {
            "summary": "Path used to manage the list of todos.",
            "description": "The REST endpoint/path used to list and create zero or more `ToDo` entities.  This path contains a `GET` and `POST` operation to perform the list and create tasks, respectively.",
            "get": {
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ToDo"
                                    }
                                }
                            }
                        },
                        "description": "Successful response - returns an array of `ToDo` entities."
                    }
                },
                "operationId": "getToDos",
                "summary": "List All ToDos",
                "description": "Gets a list of all `ToDo` entities."
            },
            "post": {
                "requestBody": {
                    "description": "A new `ToDo` to be created.",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ToDo"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Successful response."
                    }
                },
                "operationId": "createToDo",
                "summary": "Create a ToDo",
                "description": "Creates a new instance of a `ToDo`."
            }
        },
        "/todos/{todoId}": {
            "summary": "Path used to manage a single ToDo.",
            "description": "The REST endpoint/path used to get, update, and delete single instances of an `ToDo`.  This path contains `GET`, `PUT`, and `DELETE` operations used to perform the get, update, and delete tasks, respectively.",
            "get": {
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ToDo"
                                }
                            }
                        },
                        "description": "Successful response - returns a single `ToDo`."
                    }
                },
                "operationId": "getToDo",
                "summary": "Get a ToDo",
                "description": "Gets the details of a single instance of a `ToDo`."
            },
            "put": {
                "requestBody": {
                    "description": "Updated `ToDo` information.",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ToDo"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "202": {
                        "description": "Successful response."
                    }
                },
                "operationId": "updateToDo",
                "summary": "Update a ToDo",
                "description": "Updates an existing `ToDo`."
            },
            "delete": {
                "responses": {
                    "204": {
                        "description": "Successful response."
                    }
                },
                "operationId": "deleteToDo",
                "summary": "Delete a ToDo",
                "description": "Deletes an existing `ToDo`."
            },
            "parameters": [
                {
                    "name": "todoId",
                    "description": "A unique identifier for a `ToDo`.",
                    "schema": {
                        "type": "string"
                    },
                    "in": "path",
                    "required": true
                }
            ]
        }
    },
    "components": {
        "schemas": {
            "ToDo": {
                "title": "Todo",
                "description": "",
                "type": "object",
                "properties": {
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "complete": {
                        "type": "boolean"
                    },
                    "dueDate": {
                        "format": "date-time",
                        "type": "string"
                    }
                },
                "example": {
                    "title": "The title of the Todo",
                    "description": "Details for the Todo",
                    "complete": false,
                    "dueDate": "2019-12-22T09:00:00Z"
                }
            }
        }
    }
}